feat(core): injectableRecord PoC
